refactor : refactor influxSeq

+++ java-module/mc-o11y-manager/src/main/java/com/mcmp/o11ymanager/entity/TargetEntity.java
@@ -40,7 +40,7 @@ public class TargetEntity {
   @JoinColumn(name = "influxdb_id")
   private InfluxEntity influxDb;
 
-  private int influxSeq;
+  private Long influxSeq;
 
   @Column(nullable = false)
   private String name;
